finally done
this was weeks in the making, javascript is wonderful but the timing of things can be a royal pain.

very happy i was able to diagnose and troubleshoot this one

basically made another dictionary which contained the changes, and then we were able to propagate and update, while not yet applying the changes. Then, we could propagate after we updated the screen.

yay